Riptide Statistics
Provides an example of how this program can be applied on Roblox items. In this case, RiptideCount was used to count the amount of owners for the Riptide gear item. Using these statistics, one can prove that the Percy Jackson: Sea of Monsters event game temporarily re-opened on December 2, 2013 to commemorate the release of Percy Jackson: Sea of Monsters on Blu-Ray.